+++ src/components/Field.tsx
@@ -132,6 +132,13 @@ export function Field<
     throw new Error("Could not connect to form context. Check `form` prop or wrap component with a Zod Forms `<Form />` component.");
   }
 
+  // Get form data
+  const {
+    data,
+    schema,
+    validation,
+  } = form;
+
   // If the field is not hidden, make it focusable with keyboard shortcuts
   if (type !== "hidden" && !("tabIndex" in rest)) {
     rest.tabIndex ||= 0;
@@ -144,7 +151,7 @@ export function Field<
 
     // Get the default value from the current action data
     const defaultValue = path.pickFrom(
-      form?.data?.validation?.data
+      data?.validation?.data
     );
 
     // If the default value exists
@@ -154,7 +161,7 @@ export function Field<
     }
 
     // Get the shape of the field schema
-    let shape = path.toSchema(form.schema);
+    let shape = path.toSchema(schema);
 
     // If we found the shape for this field
     if (shape) {
@@ -267,6 +274,23 @@ export function Field<
           Object.assign(rest, {
             // Remove the first `/` and trailing `/` and flags as HTML does not allow them
             pattern: String(check._zod.def.pattern).substring(1).replace(/\/([a-z]+)?$/i, ""),
+          });
+        }
+      }
+
+      // If there are validation issues
+      if (validation?.error?.issues?.length) {
+        // Get the field issues
+        const issues = validation.error.issues
+          .filter(
+            issue =>
+              path.is(issue.path)
+          );
+
+        // If the field has issues
+        if (issues.length > 0) {
+          Object.assign(rest, {
+            "aria-invalid": true,
           });
         }
       }