+++ src/components/Message.tsx
@@ -3,11 +3,14 @@ import type { Paths } from "type-fest";
 import { z } from "zod";
 import { ZodFormContext } from "../context/FormContext";
 import { ZodFormsContext } from "../context/FormsContext";
+import type { HandleZodFormMessage } from "../hooks/handleZodForm";
+import { Path } from "../utils/Path";
 
 /**
  * Props for the Message component
  */
 export interface ZodFormMessageProps<
+  PayloadType,
   SchemaType extends z.ZodInterface<any>,
   FieldPath extends Paths<z.infer<SchemaType>, { bracketNotation: true; }>,
 > extends Omit<
@@ -31,6 +34,16 @@ export interface ZodFormMessageProps<
    */
   children?: (props: AllHTMLAttributes<HTMLElement> & {
     /**
+     * Issues relating to this field
+     */
+    issues?: z.core.$ZodIssue[];
+
+    /**
+     * The message to display
+     */
+    message?: HandleZodFormMessage<SchemaType, PayloadType>;
+
+    /**
      * Form validation result
      */
     validation?: z.ZodSafeParseResult<z.infer<SchemaType>>;
@@ -43,10 +56,11 @@ export interface ZodFormMessageProps<
 }
 
 export function Message<
+  PayloadType,
   SchemaType extends z.ZodInterface<any>,
   FieldPath extends Paths<z.infer<SchemaType>, { bracketNotation: true; }>,
 > (
-  props: ZodFormMessageProps<SchemaType, FieldPath>
+  props: ZodFormMessageProps<PayloadType, SchemaType, FieldPath>
 ) {
   let {
     as: Element = "div",
@@ -84,18 +98,91 @@ export function Message<
     throw new Error("Could not connect to form context. Check `form` prop or wrap component with a Zod Forms `<Form />` component.");
   }
 
+  // Get validation result from context
+  const {
+    data,
+    schema,
+    validation,
+  } = form;
+
+  // If a field name is not set, display the data message
+  if (!name) {
+    // If there is not a message
+    if (!data?.message && !data?.status) {
+      return undefined;
+    }
+
+    return (
+      children
+        ? (
+          <>
+            { children({
+              message: data,
+              ...rest,
+            }) }
+          </>
+        )
+        : (
+          <Element
+            data-status={ data.status }
+            title={ data.message }
+            { ...rest }>
+            <p>
+              { data.message }
+            </p>
+          </Element>
+        )
+    );
+  }
+
+  // If validation is not set for this field
+  if (!validation || validation.success || validation.error.issues.length === 0) {
+    return undefined;
+  }
+
+  // Get the field path
+  const fieldPath = new Path(name);
+
+  // Get the field issues
+  const issues = validation.error.issues
+    .filter(
+      issue =>
+        fieldPath.is(issue.path)
+    );
+
+  // If there are no issues for this field
+  if (issues.length === 0) {
+    return undefined;
+  }
+
   return (
     children
       ? (
         <>
-          { children({ ...rest }) }
+          { children({ ...rest, issues, validation }) }
         </>
       )
       : (
         <Element { ...rest }>
-          <p>
-            { JSON.stringify({}) }
-          </p>
+          <ul>
+            { issues.map(
+              issue => {
+                const fieldSchema: z.ZodType<any> | undefined = schema && fieldPath.toSchema(schema);
+
+                return (
+                  <li data-issue-code={ issue.code }>
+                    <strong>
+                      { fieldSchema?.meta()?.description || fieldPath.toString() }
+                    </strong>
+                    <span>
+                      { issue.message }
+                    </span>
+                  </li>
+                );
+              }
+            )
+              .filter(Boolean) }
+          </ul>
         </Element>
       )
   );